% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swReg.R
\name{plot.swReg}
\alias{plot.swReg}
\title{Plot forward stagewise regression path}
\usage{
\method{plot}{swReg}(object, legend = TRUE)
}
\arguments{
\item{object}{an R object resulting from application of \code{swReg}}

\item{legend}{logical. Should a legend be printed? Defaults to TRUE}
}
\value{
A plot with iteration numbers on the x-axis and coefficient values on the y-axis,
}
\description{
\code{plot.swReg} plots non-zero coefficients for each iterations of forward stagewise regression.
}

