% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swReg.R
\name{predict.swReg}
\alias{predict.swReg}
\title{Get predictions for a given swReg model}
\usage{
\method{predict}{swReg}(object, newdata = object$data$X, ...)
}
\arguments{
\item{object}{an R object resulting from application of \code{swReg}}

\item{newdata}{a dataframe or matrix with predictor variable values.}

\item{...}{currently not used.}
}
\value{
A vector of predictions for \code{newdata}.
}
\description{
\code{predict.swReg} returns predictions for a given swReg model
}
\examples{
## Example using Boston Housing data:
 library(MASS)
 X <- as.matrix(Boston[1:400,-14])
 Y <- Boston$medv[1:400]
 swReg1 <- swReg(X, Y, stepsize = .01)
 swReg2 <- swReg(X, Y, stepsize = .01, standardizeY = FALSE)
 plot.swReg(swReg1, newdata = Boston[401:502,-14])
 plot.swReg(swReg2, newdata = Boston[401:502,-14])
}
