% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swReg.R
\name{swReg}
\alias{swReg}
\title{Perform forward stagewise regression}
\usage{
swReg(X, Y, stepsize = 0.1, threshold = 0.1)
}
\arguments{
\item{X}{Matrix of predictor variables.}

\item{Y}{Outcome variable (vector).}

\item{stepsize}{Value with which the standardized coefficients are updated in each stage (iteration).}

\item{threshold}{Convergence criterion: When none of the predictors has a correlation greater than or equal to the threshold, the algorithm converges.}
}
\value{
The function returns a list with the following elements:
unstandardized.coef = bUnstand, 
standardized.coef = b,
iteration = iteration, 
stepsize = stepsize, 
threshold = threshold, 
coef.path = dataframe with standardized coefficient values for each predictor variable, at each stage or iteration of the algorithm
data = list with original data (X andY)
}
\description{
\code{swReg} performs forward stagewise regression.
}
\details{
The function performs forward stagewise regression, as described in Efron, B., Hastie, T., Johnstone, I., & Tibshirani, R. (2004). Least angle regression. The Annals of Statistics, 32(2), 407-499.
}
\examples{
## Example using Boston housing data:
 library(MASS)
 X <- as.matrix(Boston[,-14])
 Y <- Boston$medv
   
 ## Run forward stagewise regression:
 swReg.2 <- swReg(X, Y, st = .2, thr = .2)
 swReg.1 <- swReg(X, Y, st = .1, thr = .1)
 swReg.01 <- swReg(X, Y, st = .01, thr = .01)
 swReg.001 <- swReg(X, Y, st = .001, thr = .001)
   
 ## Plot coefficient paths:
 par(mfrow=c(2,2))
 plot.swReg(swReg.2)
 plot.swReg(swReg.1)
 plot.swReg(swReg.01)
 plot.swReg(swReg.001)
 @export
}

