% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swReg.R
\name{swReg.xval}
\alias{swReg.xval}
\title{Get k-fold cross validated predictions for a given swReg model}
\usage{
swReg.xval(object, k = 10)
}
\arguments{
\item{object}{an R object resulting from application of \code{swReg}}

\item{k}{integer. Number of folds to be used.}
}
\value{
A list with iteration numbers on the x-axis and coefficient values on the y-axis,
}
\description{
\code{swReg.xval} returns cross validated predictions for the swReg model specified.
}
