% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swReg.R
\name{plot.swReg}
\alias{plot.swReg}
\title{Plot forward stagewise regression path}
\usage{
\method{plot}{swReg}(x, legend = TRUE, ...)
}
\arguments{
\item{x}{an R object resulting from application of \code{swReg}}

\item{legend}{logical. Should a legend be printed? Defaults to TRUE}

\item{...}{not used}
}
\value{
A plot with iteration numbers on the x-axis and coefficient values on the y-axis,
}
\description{
\code{plot.swReg} plots non-zero coefficients for each iterations of forward stagewise regression.
}
\examples{
## Example using Boston Housing data:
 library(MASS)
 X <- as.matrix(Boston[,-14])
 Y <- Boston$medv
 swReg1 <- swReg(X, Y, stepsize = .01)
 swReg2 <- swReg(X, Y, stepsize = .01, standardizeY = FALSE)
   
 ## Plot coefficient paths:
 par(mfrow=c(1,2))
 plot.swReg(swReg1)
 plot.swReg(swReg2)
 
}
